refactor(threads): useThreadsReducer 一期零风险外移
- 3728→3495 行：threadReducerEqualityGuards 146 + threadReducerProviderBinding 84
- applyCompleteAgentMessageToState 依赖面 22 项超零风险界，回退留二期（记录于 PR）
- __profile / threadReducerTestProjection 锚点未动；逐字等价经 diff 自证
- baseline 重生成 + 新文件登记豁免
验证：tsc 零 error；reducer 测试 83/84（1 失败基线既有）；gate 绿
